remove unnecessary `Arc` and refactor
