+++ exchange/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod adapter;
-mod client;
 pub mod connect;
 pub mod depth;
 pub mod fetcher;
